requirements.txt
amqp==5.0.9
asgiref==3.5.0
billiard==3.6.4.0
celery==5.2.3
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.12
click==8.0.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
cryptography==36.0.1
cssselect2==0.4.1
defusedxml==0.7.1
Deprecated==1.2.13
Django==4.0.2
django-allauth==0.48.0
django-ckeditor==6.2.0
django-filter==21.1
django-js-asset==2.0.0
easy-thumbnails==2.8.1
idna==3.3
kombu==5.2.3
lxml==4.7.1
oauthlib==3.2.0
packaging==21.3
Pillow==9.0.1
prompt-toolkit==3.0.28
pycparser==2.21
PyJWT==2.3.0
pyparsing==3.0.7
python3-openid==3.2.0
pytz==2021.3
redis==4.1.4
reportlab==3.6.6
requests==2.27.1
requests-oauthlib==1.3.1
six==1.16.0
sqlparse==0.4.2
svglib==1.2.1
tinycss2==1.1.1
urllib3==1.26.8
vine==5.0.0
wcwidth==0.2.5
webencodings==0.5.1
wrapt==1.13.3
